refactor: rename 'src/types/store' from '.../state'
